+++ docs/CHANGELOG.rst
@@ -46,6 +46,7 @@ Fixed
 - Fix Cuffnorm test
 - Fix tests using deprecated flow_collection
 - Fix HTSeq CPM test
+- Handle integer expression values in hierarchical clustering
 
 
 =================